+++ src/widgets/home/index.ts
@@ -1,0 +1,5 @@
+export { HeroSection } from "./HeroSection";
+export { FeaturesSection } from "./FeaturesSection";
+export { EstimateSection } from "./EstimateSection";
+export { ReviewsSection } from "./ReviewsSection";
+export { CtaSection } from "./CtaSection";
